mn: organize repos, rename files
