+++ src/dsnotificationinterface.cpp
@@ -6,10 +6,10 @@
 #include <chainparams.h>
 #include <dsnotificationinterface.h>
 #include <instantx.h>
-#include <governance.h>
-#include <masternodeman.h>
-#include <masternode-payments.h>
-#include <masternode-sync.h>
+#include <governance/governance.h>
+#include <masternode/manager.h>
+#include <masternode/payments.h>
+#include <masternode/sync.h>
 #include <privatesend.h>
 #ifdef ENABLE_WALLET
 #include <privatesend-client.h>
