+++ src/governance/governance.h
@@ -3,8 +3,8 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#ifndef DASH_GOVERNANCE_H
-#define DASH_GOVERNANCE_H
+#ifndef DASH_GOVERNANCE_GOVERNANCE_H
+#define DASH_GOVERNANCE_GOVERNANCE_H
 
 //#define ENABLE_DASH_DEBUG
 
@@ -12,9 +12,9 @@
 #include <cachemap.h>
 #include <cachemultimap.h>
 #include <chain.h>
-#include <governance-exceptions.h>
-#include <governance-object.h>
-#include <governance-vote.h>
+#include <governance/exceptions.h>
+#include <governance/object.h>
+#include <governance/vote.h>
 #include <net.h>
 #include <sync.h>
 #include <timedata.h>
@@ -461,4 +461,4 @@ private:
 
 };
 
-#endif // DASH_GOVERNANCE_H
+#endif //DASH_GOVERNANCE_GOVERNANCE_H
