+++ src/governance/validators.cpp
@@ -3,7 +3,7 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <governance-validators.h>
+#include <governance/validators.h>
 
 #include <key_io.h>
 #include <util/strencodings.h>
