+++ src/governance/vote.cpp
@@ -3,10 +3,10 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <governance-vote.h>
-#include <governance-object.h>
-#include <masternode-sync.h>
-#include <masternodeman.h>
+#include <governance/vote.h>
+#include <governance/object.h>
+#include <masternode/sync.h>
+#include <masternode/manager.h>
 #include <messagesigner.h>
 #include <util/system.h>
