+++ src/governance/votedb.cpp
@@ -3,7 +3,7 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <governance-votedb.h>
+#include <governance/votedb.h>
 
 CGovernanceObjectVoteFile::CGovernanceObjectVoteFile()
     : nMemoryVotes(0),
