+++ src/masternode/payments.h
@@ -8,7 +8,7 @@
 
 #include <core_io.h>
 #include <key.h>
-#include <masternode.h>
+#include <masternode/masternode.h>
 #include <net_processing.h>
 #include <util/strencodings.h>
 #include <util/system.h>
