+++ src/masternode/sync.cpp
@@ -3,14 +3,14 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <activemasternode.h>
 #include <checkpoints.h>
-#include <governance.h>
+#include <governance/governance.h>
 #include <validation.h>
-#include <masternode.h>
-#include <masternode-payments.h>
-#include <masternode-sync.h>
-#include <masternodeman.h>
+#include <masternode/activemasternode.h>
+#include <masternode/masternode.h>
+#include <masternode/payments.h>
+#include <masternode/sync.h>
+#include <masternode/manager.h>
 #include <netfulfilledman.h>
 #include <netmessagemaker.h>
 #include <spork.h>
