+++ src/net_processing.cpp
@@ -36,11 +36,11 @@
 
 // Dash
 #include <spork.h>
-#include <governance.h>
+#include <governance/governance.h>
 #include <instantx.h>
-#include <masternode-payments.h>
-#include <masternode-sync.h>
-#include <masternodeman.h>
+#include <masternode/payments.h>
+#include <masternode/sync.h>
+#include <masternode/manager.h>
 #ifdef ENABLE_WALLET
 #include <privatesend-client.h>
 #endif // ENABLE_WALLET
