+++ src/qt/bitcoin.cpp
@@ -31,7 +31,7 @@
 #endif
 
 // Dash
-#include <masternodeconfig.h>
+#include <masternode/config.h>
 //
 
 #include <interfaces/handler.h>
