+++ src/qt/clientmodel.cpp
@@ -26,8 +26,8 @@
 #include <warnings.h>
 
 // Dash
-#include <masternodeman.h>
-#include <masternode-sync.h>
+#include <masternode/manager.h>
+#include <masternode/sync.h>
 #include <privatesend.h>
 //
