+++ src/rpc/masternode.cpp
@@ -4,15 +4,15 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include <activemasternode.h>
 #include <init.h>
 #include <netbase.h>
 #include <key_io.h>
 #include <validation.h>
-#include <masternode-payments.h>
-#include <masternode-sync.h>
-#include <masternodeconfig.h>
-#include <masternodeman.h>
+#include <masternode/activemasternode.h>
+#include <masternode/payments.h>
+#include <masternode/sync.h>
+#include <masternode/config.h>
+#include <masternode/manager.h>
 #ifdef ENABLE_WALLET
 #include <privatesend-client.h>
 #endif // ENABLE_WALLET
