+++ src/rpc/misc.cpp
@@ -25,7 +25,7 @@
 #include <warnings.h>
 
 // Dash
-#include <masternode-sync.h>
+#include <masternode/sync.h>
 #include <spork.h>
 // FXTC BEGIN
 #include <wallet/rpcwallet.h>
